+++ source/_components/dividers/heading.blade.php
@@ -1,0 +1,11 @@
+<h3 class="flex items-center my-8">
+    <span
+        aria-hidden="true"
+        class="flex-grow bg-gray-200 rounded h-0.5"
+    ></span>
+    <span class="mx-3 text-lg font-medium">Heading</span>
+    <span
+        aria-hidden="true"
+        class="flex-grow bg-gray-200 rounded h-0.5"
+    ></span>
+</h3>